+++ app/src/hyprland_listener/types.rs
@@ -1,0 +1,30 @@
+use serde::{Deserialize, Serialize};
+
+pub enum HyprSocketType {
+    Event,
+    Command,
+}
+
+#[derive(Debug, PartialEq)]
+pub enum HyprEvent {
+    WorkspaceCreated,
+    WorkspaceChanged,
+    WorkspaceDestroyed,
+    WindowChanged,
+    Window2Changed,
+    InvalidEvent,
+    // More events will be handle in the future
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
+pub struct HyprWinInfo {
+    pub class: String,
+    pub title: String,
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq, Eq)]
+pub struct HyprWorkspaceInfo {
+    pub id: u32,
+    pub occupied: bool,
+    pub active: bool,
+}
